Add backoff to queryStream, scanStream.
A reasonable user of these streaming functions would expect them to stream
data as long as there is data (paging, already implemented), and at the rate
it can be fetched.

This commit takes the specific errors ProvisionedThroughputExceededException
and ThrottlingException, and handles them as signals to temporarily back off.

It starts retrying after 200 ms, then doubles that each time it retries
without a successful response, up to 60000 ms. Then it retries every 60000 ms.